arreglo de vectores

+++ clasePunto/punto.h
@@ -1,12 +1,12 @@
 class Point {
     private:
-        int x; int y; bool is_null;
+        float x; float y; bool is_null;
     public:
         Point();
         Point(int x, int y);
         Point(Point &p);
-         int getX() const;
-         int getY() const;
+        float getX() const;
+        float getY() const;
         void setX(int x);
         void setY(int y);
         void printPoint();
